# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UrbanWatch API - SAR and optical satellite urban growth detection
 * OpenAPI spec version: 0.1.0
 */
import type { GrowthRegion } from "./growthRegion";
import type { LandUseSummary } from "./landUseSummary";

export interface AnalysisSummary {
  totalUrbanAreaKm2: number;
  totalAreaKm2: number;
  urbanCoveragePercent: number;
  annualGrowthRatePercent: number;
  /** Urban area gained in the last 30 days */
  recentGrowthKm2: number;
  activeRegionsCount: number;
  lastUpdated: string;
  landUseSummary: LandUseSummary;
  topGrowthRegions: GrowthRegion[];
}