# lib/api-zod/src/generated/types/areaOfInterest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UrbanWatch API - SAR and optical satellite urban growth detection
 * OpenAPI spec version: 0.1.0
 */

export interface AreaOfInterest {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  minLat: number;
  maxLat: number;
  minLon: number;
  maxLon: number;
  areaKm2: number;
  createdAt: string;
  updatedAt: string;
}